+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/artifacts/transform/CollectorTransformer.kt
@@ -33,6 +33,7 @@ import kotlin.io.path.listDirectoryEntries
 
 /**
  * The artifact transformer collecting JAR files located within the IntelliJ Platform or plugin archives.
+ * @see org.jetbrains.intellij.platform.gradle.artifacts.transform.LocalIvyArtifactPathComponentMetadataRule
  */
 @DisableCachingByDefault(because = "Not worth caching")
 abstract class CollectorTransformer : TransformAction<CollectorTransformer.Parameters> {
@@ -87,11 +88,16 @@ abstract class CollectorTransformer : TransformAction<CollectorTransformer.Param
                 }
 
                 isPlugin -> {
+                    // Normally we get into here for 3rd party marketplace plugins.
+                    // https://plugins.jetbrains.com/docs/intellij/tools-intellij-platform-gradle-plugin-dependencies-extension.html#non-bundled-plugin
+                    // For other plugins, we never (usually?) get into this block, because their Ivy artifacts already
+                    // list jars, instead of pointing to a directory, see:
+                    // org.jetbrains.intellij.platform.gradle.artifacts.transform.LocalIvyArtifactPathComponentMetadataRule
                     plugin.originalFile?.let { pluginPath ->
-                        collectJars(
-                            pluginPath.resolve("lib"),
-                            pluginPath.resolve("lib/modules"),
-                        ).forEach { outputs.file(it) }
+                        val jars = collectJars(pluginPath)
+                        jars.forEach {
+                            outputs.file(it)
+                        }
                     }
                 }
 
@@ -102,11 +108,16 @@ abstract class CollectorTransformer : TransformAction<CollectorTransformer.Param
         }
     }
 
-    private fun collectJars(vararg paths: Path?) = paths
-        .mapNotNull { it?.takeIfExists() }
-        .flatMap { it.listDirectoryEntries("*.jar") }
-
     companion object {
+        internal fun collectJars(path: Path): List<Path> {
+            val libPath = path.resolve("lib")
+            val libModulesPath = libPath.resolve("modules")
+
+            return listOf(libPath, libModulesPath)
+                .mapNotNull { it.takeIfExists() }
+                .flatMap { it.listDirectoryEntries("*.jar") }
+        }
+
         internal fun register(
             dependencies: DependencyHandler,
             compileClasspathConfiguration: Configuration,