+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/IntelliJPlatformDependenciesHelper.kt
@@ -13,6 +13,7 @@ import org.gradle.api.artifacts.ProjectDependency
 import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.api.file.Directory
 import org.gradle.api.file.ProjectLayout
+import org.gradle.api.initialization.resolve.RulesMode
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ProviderFactory
@@ -39,7 +40,6 @@ import java.io.File
 import java.io.FileReader
 import java.nio.file.Path
 import java.util.*
-import kotlin.Throws
 import kotlin.io.path.*
 
 /**
@@ -59,6 +59,7 @@ class IntelliJPlatformDependenciesHelper(
     private val objects: ObjectFactory,
     private val providers: ProviderFactory,
     private val rootProjectDirectory: Path,
+    private val metadataRulesModeProvider: Provider<RulesMode>
 ) {
 
     private val log = Logger(javaClass)
@@ -687,7 +688,7 @@ class IntelliJPlatformDependenciesHelper(
                     module = type.code,
                     revision = version,
                 ),
-                publications = listOf(artifactPath.toIvyArtifact()),
+                publications = listOf(artifactPath.toAbsolutePathIvyArtifact()),
             )
         }
 
@@ -750,7 +751,7 @@ class IntelliJPlatformDependenciesHelper(
                     module = id,
                     revision = version,
                 ),
-                publications = listOf(artifactPath.toIvyArtifact()),
+                publications = artifactPath.toIvyArtifacts(metadataRulesModeProvider, platformPath.get()),
                 dependencies = plugin.collectDependencies(),
             )
         }
@@ -772,7 +773,9 @@ class IntelliJPlatformDependenciesHelper(
             .find { layout -> layout.name == id }
             .let { requireNotNull(it) { "Specified bundledModule '$id' doesn't exist." } }
         val platformPath = platformPath.get()
-        val artifactPaths = bundledModule.classPath.map { path -> platformPath.resolve(path).toIvyArtifact() }
+        val artifactPaths = bundledModule.classPath.flatMap {
+            path -> platformPath.resolve(path).toIvyArtifacts(metadataRulesModeProvider, platformPath)
+        }
         val version = baseVersion.orElse(productInfo.map { it.version }).get()
 
         writeIvyModule(Dependencies.BUNDLED_MODULE_GROUP, id, version) {
@@ -816,7 +819,7 @@ class IntelliJPlatformDependenciesHelper(
                 writeIvyModule(group, name, version) {
                     IvyModule(
                         info = IvyModule.Info(group, name, version),
-                        publications = listOf(artifactPath.toIvyArtifact()),
+                        publications = artifactPath.toIvyArtifacts(metadataRulesModeProvider, platformPath),
                         dependencies = when {
                             id in path -> emptyList()
                             else -> plugin.collectDependencies(path + id)
@@ -836,7 +839,9 @@ class IntelliJPlatformDependenciesHelper(
             .mapNotNull { layoutItems.find { layout -> layout.name == it } }
             .filterNot { it.classPath.isEmpty() }
             .map {
-                val artifactPaths = it.classPath.map { path -> platformPath.resolve(path).toIvyArtifact() }
+                val artifactPaths = it.classPath.flatMap {
+                    path -> platformPath.resolve(path).toIvyArtifacts(metadataRulesModeProvider, platformPath)
+                }
                 val group = Dependencies.BUNDLED_MODULE_GROUP
                 val name = it.name
                 val version = buildNumber
@@ -885,7 +890,7 @@ class IntelliJPlatformDependenciesHelper(
                     module = name,
                     revision = version,
                 ),
-                publications = listOf(artifactPath.toIvyArtifact()),
+                publications = listOf(artifactPath.toAbsolutePathIvyArtifact()),
             )
         }
 
@@ -926,7 +931,7 @@ class IntelliJPlatformDependenciesHelper(
                     module = name,
                     revision = version,
                 ),
-                publications = listOf(artifactPath.toIvyArtifact()),
+                publications = listOf(artifactPath.toAbsolutePathIvyArtifact()),
             )
         }
 
@@ -999,10 +1004,26 @@ class IntelliJPlatformDependenciesHelper(
             extension = "zip",
         )
 
+
     /**
      * Creates a [Provider] that holds a JetBrains Runtime version obtained using the currently used IntelliJ Platform.
      */
     internal fun obtainJetBrainsRuntimeVersion() = cachedProvider {
+        val version = tryObtainJetBrainsRuntimeVersion()
+        if (!version.isNullOrEmpty()) {
+            return@cachedProvider buildJetBrainsRuntimeVersion(version)
+        }
+
+        return@cachedProvider tryObtainBundledJetBrainsRuntimeVersion()
+    }
+
+    /**
+     * The Multi-OS Archives don't bundle JetBrains Runtime (JBR) needed to run the IDE locally, perform testing,
+     * and other crucial operations. Therefore, it is required to explicitly add a dependency on JetBrains Runtime
+     * (JBR) by adding extra jetbrainsRuntime() repository and dependency entries.
+     * On these dependencies.txt is not present.
+     */
+    private fun tryObtainJetBrainsRuntimeVersion(): String? {
         val version = platformPath.get().resolve("dependencies.txt").takeIf { it.exists() }?.let {
             FileReader(it.toFile()).use { reader ->
                 with(Properties()) {
@@ -1010,9 +1031,61 @@ class IntelliJPlatformDependenciesHelper(
                     getProperty("runtimeBuild") ?: getProperty("jdkBuild")
                 }
             }
-        } ?: throw GradleException("Could not obtain JetBrains Runtime version with the current IntelliJ Platform.")
+        }
 
-        buildJetBrainsRuntimeVersion(version)
+        return version
+    }
+
+    /**
+     * IntelliJ Platform installers are OS-specific and contain bundled JetBrains Runtime (JBR), but are limited
+     * to public releases only.
+     * Installers are always used when running the verifyPlugin task to perform the binary compatibility checks.
+     */
+    private fun tryObtainBundledJetBrainsRuntimeVersion(): String {
+        val releaseFile = with(OperatingSystem.current()) {
+            when {
+                isMacOsX -> "jbr/Contents/Home/release"
+                isLinux -> "jbr/release"
+                isWindows -> "jbr/release"
+                else -> throw GradleException("Failed to obtain platform OS for: $this")
+            }
+        }
+
+        val releaseFilePath = platformPath.get().resolve(releaseFile).takeIf { it.exists() }
+            ?: throw GradleException("Could not obtain JetBrains Runtime version with the current IntelliJ Platform.")
+
+        val properties = releaseFilePath.let {
+            FileReader(it.toFile()).use { reader ->
+                with(Properties()) {
+                    load(reader)
+                    this
+                }
+            }
+        }
+
+        // JAVA_RUNTIME_VERSION property is missing in old versions.
+        // E.g., JAVA_VERSION="17.0.6"
+        val javaVersion = properties.getProperty("JAVA_VERSION")?.replace("\"", "")
+            ?: throw GradleException("Could not obtain JetBrains Runtime version with the current IntelliJ Platform.")
+        // E.g., IMPLEMENTOR_VERSION="JBR-17.0.6+1-653.34-jcef"
+        // Very old IDE's like IU-192.7142.36 do not have this property, they have only JAVA_VERSION & IMPLEMENTOR="N/A"
+        val implementorVersion = properties.getProperty("IMPLEMENTOR_VERSION")?.replace("\"", "")
+            ?: throw GradleException("Could not obtain JetBrains Runtime version with the current IntelliJ Platform.")
+        // E.g. +1-653.34
+        val subString = implementorVersion.substringAfter(javaVersion).removeSuffix("-jcef").removeSuffix("-JCEF")
+        // E.g. 653.34-jcef
+        val buildVersion = subString.replace(Regex("^([+][0-9]+-)"), "")
+        // E.g. 17.0.6-b653.34-jcef
+        // It is actually could be also "17.0.6-b829.4"
+        // https://github.com/JetBrains/JetBrainsRuntime/releases?page=8
+        // The latest are without -
+        // https://github.com/JetBrains/JetBrainsRuntime/releases
+        val fixedVersion = "${javaVersion}b${buildVersion}"
+
+        val isJcef = implementorVersion.endsWith("-jcef") || implementorVersion.endsWith("-JCEF")
+        val variant = if (isJcef) "jcef" else null
+
+        return buildJetBrainsRuntimeVersion(fixedVersion, variant)
     }
 
     /**
@@ -1119,8 +1192,12 @@ class IntelliJPlatformDependenciesHelper(
         val variant = runtimeVariant ?: "jcef"
 
         val (jdk, build) = version.split('b').also {
-            assert(it.size == 1) {
-                "Incorrect JetBrains Runtime version: $version. Use [sdk]b[build] format, like: 21.0.3b446.1"
+            assert(it.size == 2) {
+                // It is actually could be also "17.0.6-b829.4"
+                // https://github.com/JetBrains/JetBrainsRuntime/releases?page=8
+                // The latest are without -
+                // https://github.com/JetBrains/JetBrainsRuntime/releases
+                "Incorrect JetBrains Runtime version: '$version'. Use [sdk]b[build] format, like: '21.0.3b446.1'"
             }
         }