+++ CHANGELOG.md
@@ -9,12 +9,14 @@
 ### Changed
 
 - Move `localPlatformArtifacts()` to the top of the `defaultRepositories()` list
+- Improved build performance by making the local Ivy repository an exclusive Gradle repository.
 
 ### Fixed
 
-- Fixed issue #1778 by removing a hash of the absolute artifact path appended to the end of the version string. That hash made artifact version different on different PCs and also breaks Gradle dependency locking.
+- Fixed issue #1778 by removing a hash of the absolute artifact path appended to the end of the version string. That hash made the versions different on different PCs and also breaks Gradle dependency locking.
 - Add the missing `org.jetbrains.kotlin.platform.type=jvm` attribute to the `intellijPlatformRuntimeClasspath` configuration manually as it's not inherited from the `runtimeClasspath`.
 - Fixed `Could not generate a decorated class for type PluginArtifactRepository.` when creating a custom plugin repository.
+- Fixed issue #1779 about compatibility with Gradle dependency verification. Previously it was failing with "Failed to create MD5 hash for file".
 
 ## [2.1.0]